examples/integrations/gcp/src/etl/images/bounding_box/requirements.txt
google-cloud-storage
labelbox[data]
